+++ mediaclub_laravel/database/migrations/2025_05_07_131113_create_hilo_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('hilo', function (Blueprint $table) {
             $table->integer('hilo_id', true);
             $table->integer('usuario_id')->index('usuario_id');
-            $table->integer('frame_id')->index('frame_id');
+            $table->unsignedBigInteger('frame_id')->index('frame_id');
             $table->string('titulo');
             $table->text('contenido');
             $table->dateTime('fecha_creacion');
