+++ mediaclub_laravel/database/migrations/2025_05_07_131113_create_puntuacion_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('puntuacion', function (Blueprint $table) {
             $table->integer('puntuacion_id', true);
             $table->integer('usuario_id');
-            $table->integer('frame_id')->index('frame_id');
+            $table->unsignedBigInteger('frame_id')->index('frame_id');
             $table->decimal('puntuacion', 3, 1);
             $table->dateTime('fecha')->nullable()->useCurrent();
