+++ mediaclub_laravel/database/migrations/2025_05_07_131113_create_resena_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('resena', function (Blueprint $table) {
             $table->integer('resena_id', true);
             $table->integer('usuario_id')->index('usuario_id');
-            $table->integer('frame_id')->index('frame_id');
+            $table->unsignedBigInteger('frame_id')->index('frame_id');
             $table->dateTime('fecha')->nullable()->useCurrent();
             $table->text('contenido');
             $table->boolean('spoiler')->nullable()->default(false);
